refactor: 修正 MVP 架構的 View and Presenter 的命名
MVP 用 Contract 來集中規範 View and Presenter, 命名應該使用 View and Presenter
MainActivity -> View
MainPresenter -> Presenter
